fix: connecting to evm wallet for telos bridge
