+++ src/pages/transfer/managers/telosEvmBridge.ts
@@ -2,10 +2,11 @@ import { get } from "svelte/store";
 
 import { currentAccountBalance } from "~/store";
 import { TransferManager } from "./transferManager";
-import { Transfer } from "~/abi-types";
+import { TelosEvmOpenWallet, Transfer } from "~/abi-types";
 import { Asset, Name } from "anchor-link";
 import { updateActiveAccount } from "~/stores/account-provider";
 import { updateEvmBalance } from "~/stores/balances-provider";
+import { getTelosEvmAccount } from "~/lib/evm";
 
 export class TelosEvmBridge extends TransferManager {
     static from = "telos"
@@ -13,25 +14,29 @@ export class TelosEvmBridge extends TransferManager {
     static to = "evm"
     static toDisplayString = "TLOS (EVM)"
     static supportedChains = ["telos"]
-    static evmRequired = false;
+    static evmRequired = true;
 
     get fromAddress() {
         return String(this.nativeSession.auth.actor)
     }
 
     get toAddress() {
-        return this.evmSession.address || 'Address will be created after first transfer'
+        return this.evmSession.address
     }
 
     async transfer(amount: string) {
+        const telosEvmAccount = await getTelosEvmAccount(this.nativeSession.auth.actor)
+
+        if (!telosEvmAccount) {
+            this.createNewTelosEvmAccount()
+        }
+
         const action = Transfer.from({
             from: this.nativeSession.auth.actor,
             to: 'eosio.evm',
             quantity: String(Asset.fromFloat(Number(amount), '4,TLOS')),
             memo: this.evmSession.address || '',
         })
-
-        console.log({auth: this.nativeSession.auth})
 
         return this.nativeSession.transact({
             action: {
@@ -56,5 +61,22 @@ export class TelosEvmBridge extends TransferManager {
             updateActiveAccount(),
             updateEvmBalance(),
         ])
+    }
+
+    createNewTelosEvmAccount() {
+        const action = TelosEvmOpenWallet.from({
+            account: this.nativeSession.auth.actor,
+            address: this.evmSession.address,
+        })
+
+        this.nativeSession.transact({
+            action: {
+                authorization: [this.nativeSession.auth],
+                account: Name.from('eosio.evm'),
+                name: Name.from('openwallet'),
+                data: action,
+            },
+        })
+
     }
 }
